requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.6.1
docker==7.1.0
eventlet==0.36.1
executing==2.0.1
finnhub-python==2.4.20
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-marshmallow==0.14.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
frozendict==2.4.4
frozenlist==1.4.1
gast==0.6.0
gevent==24.2.1
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
iknl-flasgger==0.9.2.post1
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.4.1
libclang==18.1.1
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==2.16.3
marshmallow-sqlalchemy==0.15.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
multidict==6.0.5
multitasking==0.0.11
namex==0.0.8
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
peewee==3.17.5
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
questionary==1.10.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
schema==0.7.7
scikit-learn==1.5.0
scipy==1.14.0
setuptools==70.2.0
simple-websocket==1.0.0
six==1.16.0
socketIO-client==0.7.2
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
TA-Lib==0.4.32
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
textblob==0.18.0.post0
textmagic==2.0.3
threadpoolctl==3.5.0
tqdm==4.66.4
traitlets==5.14.3
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.40
zope.event==5.0
zope.interface==6.4.post2
